pyproject.toml
[project]
name = "ai_voice_agent"
version = "0.1.0"
description = "A project using Whisper, TTS, LiveKit plugins, and DeepSpeed"
authors = [
  { name="Your Name", email="you@example.com" }
]
dependencies = [
  "TTS==0.22.0",
  "deepspeed==0.16.8",
  "livekit-agents",
  "livekit-plugins-turn-detector",
  "livekit-plugins-noise-cancellation",
  "livekit-plugins-silero",
  "huggingface_hub",
  "dot",
  "python-dotenv",
  "whisper @ git+https://github.com/openai/whisper.git"
]

requires-python = ">=3.8"
